renamed types -> models to fix minor bug
